requirements.txt
beautifulsoup4
python_telegram_bot==20.2
requests==2.28.2
SQLAlchemy==2.0.5.post1
lxml
# python-dotenv